+++ idp_server/rest.py
@@ -188,6 +188,10 @@ class meta(Resource):
                 args = parser.parse_args()
                 idp = idpOf(args['code'])
                 state = State.make(idp, "{}", "{}")
+                # ensure the stateful solvers are initialized
+                _ = state.solver
+                _ = state.optimize_solver
+                _ = state.explain_solver
                 if not state.idp.display.manualRelevance:
                     get_relevant_questions(state)
                 out = metaJSON(state)
@@ -226,16 +230,13 @@ class eval(Resource):
                 parser.add_argument('method', type=str, help='Method to execute')
                 parser.add_argument('code', type=str, help='Code')
                 parser.add_argument('active', type=str, help='Three-valued structure')
-                parser.add_argument('previous_active', type=str, help='Previous input by user')
-                # parser.add_argument('expanded', type=str, action='append', help='list of expanded symbols')
+                parser.add_argument('previous_active', type=str, help='Previously propagated assignment')
                 parser.add_argument('symbol', type=str, help='Symbol to explain or optimize')
                 parser.add_argument('value', type=str, help='Value to explain')
                 parser.add_argument('field', type=str, help='Applied Symbol whose range must be determined')
                 parser.add_argument('minimize', type=bool, help='True -> minimize ; False -> maximize')
                 parser.add_argument('with_relevance', type=bool, help='compute relevance if true')
                 args = parser.parse_args()
-                #print(args)
-                # expanded = tuple([]) if args['expanded'] is None else tuple(args['expanded'])
                 method = args['method']
                 out = {}
                 if method == "checkCode":
@@ -259,10 +260,10 @@ class eval(Resource):
                     out = {"result": "ok"}
                 else:
                     state = State.make(idpOf(args['code']),
-                                    args.get('previous_active', None),
-                                    args['active'])
+                                       args['previous_active'],
+                                       args['active'])
 
-                    if not state.propagate_success:
+                    if not state.satisfied:
                         out = explain(state)
                     elif method == "propagate":
                         if args.with_relevance:
@@ -275,6 +276,7 @@ class eval(Resource):
                         out = Output(state).fill(state)
                     elif method == "modelexpand":
                         generator = state.expand(max=1, timeout=0, complete=False)
+                        # TODO: this copying is not needed?
                         out = copy(state)
                         out.assignments = out.assignments.copy(shallow=True)
                         out.assignments = list(generator)[0]
@@ -282,6 +284,7 @@ class eval(Resource):
                     elif method == "explain":
                         out = explain(state, args['value'])
                     elif method == "minimize":
+                        # TODO: this copying is not needed?
                         out = copy(state)
                         out.assignments = out.assignments.copy(shallow=True)
                         out = out.optimize(args['symbol'], args['minimize'])